Type changes for "Time" to long or double

+++ Libraries/Core/Models/Eft/Common/LocationBase.cs
@@ -274,7 +274,7 @@ public record LocationBase
     public LocationEvents? Events { get; set; }
 
     [JsonPropertyName("exit_access_time")]
-    public int? ExitAccessTime { get; set; }
+    public double? ExitAccessTime { get; set; }
 
     [JsonPropertyName("ForceOnlineRaidInPVE")]
     public bool? ForceOnlineRaidInPVE { get; set; }
@@ -286,7 +286,7 @@ public record LocationBase
     public int? ExitCount { get; set; }
 
     [JsonPropertyName("exit_time")]
-    public int? ExitTime { get; set; }
+    public double? ExitTime { get; set; }
 
     [JsonPropertyName("exits")]
     public List<Exit>? Exits { get; set; }
@@ -361,7 +361,7 @@ public record Transit
     public string? Target { get; set; }
 
     [JsonPropertyName("time")]
-    public int? Time { get; set; }
+    public long? Time { get; set; }
 }
 
 public record NonWaveGroupScenario
@@ -628,7 +628,7 @@ public record MinPlayerWaitTime
     public int? MinPlayers { get; set; }
 
     [JsonPropertyName("time")]
-    public int? Time { get; set; }
+    public long? Time { get; set; }
 }
 
 public record Preview
